pro mode

+++ server/public/assets/scripts/script.js
@@ -40,21 +40,23 @@ function receiveData () {
 function appendDom(response){
   $('.task').empty();
   response.forEach(function(task){
-    $('.task').append('<div class= "'+task.task_name+'"></div>');
-    var $el = $('.task').children().last();
-    $el.append('<h2>' + "Task content: " + task.task_name + '</h2>');
-    $el.append('<h2>' + task.task_status + '</h2>');
-
     if (task.task_status == "completed"){
-
-      $('.'+task.task_name).addClass("highlight");
-
+      $('.task').append('<div class= "'+task.task_name+'"></div>');
+      var $el = $('.task').children().last();
+      $el.append('<h2>' + "Task content: " + task.task_name + '</h2>');
+      $el.append('<h2>' + task.task_status + '</h2>');
+      $('.'+task.task_name).addClass("fade");
     }else{
-    $el.append('<button class = "complete"> complete </button>');
-    $el.append('<button class = "delete"> delete </button>');
-    var eachTask = task.task_name;
-    $el.find('.complete').data("taskCompleted", eachTask);
-    $el.find('.delete').data("taskDeleted", eachTask);
+      $('.task').prepend('<div class= "'+task.task_name+'"></div>');
+      var $el = $('.task').children().first();
+      $el.append('<h2>' + "Task content: " + task.task_name + '</h2>');
+      $el.append('<h2>' + task.task_status + '</h2>');
+      $('.'+task.task_name).addClass("highlight");
+      $el.append('<button class = "complete"> complete </button>');
+      $el.append('<button class = "delete"> delete </button>');
+      var eachTask = task.task_name;
+      $el.find('.complete').data("taskCompleted", eachTask);
+      $el.find('.delete').data("taskDeleted", eachTask);
     }
 
   })
@@ -77,11 +79,16 @@ function updateData(newData){
 
 
 function handleDelete (){
-console.log("it works");
 deletedTask.task = $(this).data("taskDeleted");
-alert("Are you sure?");
-$(this).parent().remove();
-deleteData(deletedTask);
+var txt;
+var r = confirm("Are you sure you want delete this task?");
+if (r == true) {
+  txt = "You deleted the task!";
+  $(this).parent().remove();
+  deleteData(deletedTask);
+}else {
+  txt = "Cancel delete!";
+}
 }
 
 function deleteData(deletedData){
